+++ decidim-core/spec/features/notifications_spec.rb
@@ -82,6 +82,7 @@ describe "Notifications", type: :feature do
     context "when setting all notifications as read" do
       it "hides all notifications from the page" do
         click_link "Mark all as read"
+        expect(page).not_to have_selector("#notifications-list")
         expect(page).to have_content("No notifications yet")
       end
     end