+++ decidim-core/spec/models/decidim/scope_spec.rb
@@ -4,10 +4,17 @@ require "spec_helper"
 
 module Decidim
   describe Scope do
-    subject { scope }
+    subject(:scope) { build(:scope, parent: parent) }
 
     let(:parent) { nil }
-    let(:scope) { build(:scope, parent: parent) }
+
+    describe "has an association for children scopes" do
+      subject(:scope_children) { scope.children }
+
+      let(:scopes) { create_list(:scope, 2, parent: scope) }
+
+      it { is_expected.to contain_exactly(*scopes) }
+    end
 
     context "when it is valid" do
       it { is_expected.to be_valid }
@@ -50,7 +57,8 @@ module Decidim
     end
 
     describe "cycles validation" do
-      let(:scope) { create(:scope) }
+      subject(:scope) { create(:scope) }
+
       let(:subscope) { create(:subscope, parent: scope) }
       let(:subsubscope) { create(:subscope, parent: subscope) }