+++ decidim-meetings/spec/shared/manage_registrations_examples.rb
@@ -93,9 +93,7 @@ shared_examples "manage registrations" do
           fill_in_meeting_registration_invite name: registered_user.name, email: registered_user.email
         end
 
-        logout :user
-
-        login_as user, scope: :user
+        relogin_as user
 
         visit last_email_link